+++ models/k2_horizon/__init__.py
@@ -1,0 +1,1 @@
+"""K2-Horizon model support."""